+++ changelog.md
@@ -1,5 +1,9 @@
 # 1.12.2
 
+## 0.11
+- Added Input blacklist for items from other mods that require their own EntityItem logic.
+- Added config option to disable said blacklist.
+
 ## 0.10
 - Merged custom entities so that one specific input can do multiple types of crafting instead of picking the first matching recipe.
 - Added Fire Crafting.